Switches SPI and I2C to use enums in the HAL for ports
Closes #397

+++ wpilibj/src/athena/java/edu/wpi/first/wpilibj/SPI.java
@@ -32,7 +32,7 @@ public class SPI extends SensorBase {
 
   private static int devices = 0;
 
-  private byte m_port;
+  private int m_port;
   private int m_bitOrder;
   private int m_clockPolarity;
   private int m_dataOnTrailing;
